Address upstream PR review (Goykhman): 5 fixes + tighten lint
- abi.py: drop unused `from llvmlite import ir`; replace module-level
  prose docstring with a one-liner.
- abi.py / call.py: move `_call_lib_func_byval` + `_emit_byval_call`
  to call.py (sibling intrinsic to `_call_lib_func`); abi.py keeps
  only the platform / classification / sizing primitives.
- docs/numbox.core.bindings.rst: add an "ABI dispatch" section with
  the prose previously in abi.py's docstring + an automodule entry
  for `numbox.core.bindings.abi`.
- test/core/test_abi.py: drop the retired-symbol absence loop (no
  long-term value); update the present-symbol asserts for the move;
  drop redundant `_c` / `_math` side-effect imports — `_call_lib_func`
  already triggers the parent package's `__init__.py` which loads them.
- .github/workflows/numbox_ci.yml: add F401 to the strict `--select`
  so unused imports fail the build (would have caught the dead
  `from llvmlite import ir`).
