+++ .github/workflows/numbox_ci.yml
@@ -44,8 +44,8 @@ jobs:
         pip install -e .
     - name: Lint with flake8
       run: |
-        # stop the build if there are Python syntax errors or undefined names
-        flake8 . --count --exclude=test/core/random_image_ref.py --select=E9,F63,F7,F82 --show-source --statistics
+        # stop the build if there are Python syntax errors, undefined names, or unused imports
+        flake8 . --count --exclude=test/core/random_image_ref.py --select=E9,F63,F7,F82,F401 --show-source --statistics
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: Test with pytest
