+++ docs/numbox.core.bindings.rst
@@ -16,8 +16,38 @@ Analogous technique can be expanded as needed for the user custom code.
 
 .. [#f1] See `numbsql <https://github.com/cpcloud/numbsql>`_ for previous work on jit-wrapping FFI imported functions.
 
+ABI dispatch
+++++++++++++
+
+LLVM's JIT treats ABI lowering as a frontend responsibility — it won't insert the right calling convention
+for struct args/returns by itself. ``numbox.core.bindings.call._call_lib_func`` dispatches per platform and
+per struct shape, using primitives from ``numbox.core.bindings.abi`` (platform identification via
+``_current_platform``, struct-shape classification via ``_classify``, struct-size measurement via
+``_struct_bytes``). The two ABI families that matter:
+
+- **Windows x64** — passes aggregates >8 bytes via caller-allocated pointers and returns them via ``sret``;
+  sizes 1/2/4/8 go directly in registers.
+- **SysV x86-64 / AAPCS64** — pass and return ≤16-byte aggregates directly in GP registers; on SysV x86-64,
+  >16-byte by-value args use a ``byval`` + ``optnone`` + ``noinline`` idiom so the LLVM optimizer doesn't
+  elide the caller-side stack copy before the callee reads it.
+
+References:
+
+- `llvmlite#300 <https://github.com/numba/llvmlite/issues/300#issuecomment-327235846>`_
+- `llvm-project#85417 <https://github.com/llvm/llvm-project/issues/85417>`_
+- `Windows x64 calling convention <https://learn.microsoft.com/en-us/cpp/build/x64-calling-convention>`_
+- `AAPCS64 <https://github.com/ARM-software/abi-aa/blob/main/aapcs64/aapcs64.rst>`_
+
 Modules
 ++++++++
+
+numbox.core.bindings.abi
+------------------------
+
+.. automodule:: numbox.core.bindings.abi
+   :members:
+   :show-inheritance:
+   :undoc-members:
 
 numbox.core.bindings._c
 -----------------------
